% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlts.R
\name{lpx}
\alias{lpx}
\title{Utility function}
\usage{
lpx(x, nn = 0, h = 0, adpen = 0, deg = 2, acri = "none",
  scale = FALSE, style = "none")
}
\arguments{
\item{x}{\dots{}}

\item{nn}{\dots{}}

\item{h}{\dots{}}

\item{adpen}{\dots{}}

\item{deg}{\dots{}}

\item{acri}{\dots{}}

\item{scale}{\dots{}}

\item{style}{\dots{}}
}
\description{
hack to make ll.order work with locfit >1.5. not to be called by the user.
}
\details{
not to be called by the user.
}
\author{
Catherine Loader
}
\keyword{misc}
