% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plodia}
\alias{plodia}
\title{Time series of Meal Moth abundance}
\format{A vector containing 55 values}
\usage{
plodia
}
\description{
This is replicate 3 in Bjornstad et al. (1998).
}
\references{
Bjornstad, O. N., M. Begon, N. C. Stenseth, W. Falck, S. M.
Sait, and D. J. Thompson. 1998. Population dynamics of the Indian meal moth:
demographic stochasticity and delayed regulatory mechanisms.  Journal of
Animal Ecology 67:110-126. https://doi.org/10.1046/j.1365-2656.1998.00168.x
}
\keyword{datasets}
