% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlts.R
\name{mkx}
\alias{mkx}
\title{Utility function}
\usage{
mkx(x, lags)
}
\arguments{
\item{x}{A univariate time series.}

\item{lags}{The vector of time lags.}
}
\value{
A matrix of lagged abundances. The last column is the current
}
\description{
A function to create matrix of lagged time series.  Called by various
functions.
}
\details{
If lags is \code{c(1,4)}, say, then the function returns a matrix that
consist of columns x(t-1), x(t-4), x(t).
}
\references{
Lall, U. & Sharma, A. (1996) A nearest neighbor
bootstrap for time series resampling. Water Resources Research, 32, 679-693. https://doi.org/10.1029/95wr02966
}
\author{
Upmanu Lall
}
\keyword{misc}
