% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlts.R
\name{plot.contingency.periodogram}
\alias{plot.contingency.periodogram}
\title{Plot contingency periodograms}
\usage{
\method{plot}{contingency.periodogram}(x, ...)
}
\arguments{
\item{x}{an object of class "contingency.periodogram", usually, as a result
of a call to \code{contingency.periodogram}.}

\item{...}{generic plot arguments.}
}
\value{
A contingency periodogram is plotted. The line represents the
critical value based on the chi-squared test (95\%).
}
\description{
`plot' method for "contingency.periodogram" class object.
}
\seealso{
\code{\link{contingency.periodogram}}
}
\keyword{ts}
