% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlts.R
\name{plot.lin.order}
\alias{plot.lin.order}
\title{Plot linear cross-validation for time-series order}
\usage{
\method{plot}{lin.order}(x, ...)
}
\arguments{
\item{x}{an object of class "lin.order", usually, as a result of a call to
\code{lin.order.cls} or \code{lin.order.mle}.}

\item{...}{generic plot arguments.}
}
\value{
A xy-plot of order against cross-validation error is produced.
}
\description{
`plot' method for class "lin.order".
}
\seealso{
\code{\link{lin.order.cls}}
}
\keyword{ts}
