% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlts.R
\name{plot.ll.order}
\alias{plot.ll.order}
\title{Plot nonparametric cross-validation for time-series order}
\usage{
\method{plot}{ll.order}(x, ...)
}
\arguments{
\item{x}{an object of class "ll.order", usually, as a result of a call to
\code{ll.order}.}

\item{...}{generic plot arguments.}
}
\value{
A xy-plot of minimum cross-validation error against order is
produced.
}
\description{
`plot' method for class "ll.order".
}
\details{
See \code{\link{ll.order}} for details.
}
\seealso{
\code{\link{ll.order}}
}
\keyword{ts}
