% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlts.R
\name{plot.lomb}
\alias{plot.lomb}
\title{Plot Lomb periodograms}
\usage{
\method{plot}{lomb}(x, ...)
}
\arguments{
\item{x}{an object of class "lomb", usually, as a result of a call to
\code{spec.lomb}.}

\item{...}{generic plot arguments.}
}
\value{
A Lomb periodogram is composed of a xy-plot of amplitude against
frequency.
}
\description{
`plot' method for objects of class "lomb".
}
\seealso{
\code{\link{spec.lomb}}
}
\keyword{ts}
