% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlts.R
\name{plot.ppll}
\alias{plot.ppll}
\title{Plot function for prediction profile objects}
\usage{
\method{plot}{ppll}(x, ...)
}
\arguments{
\item{x}{an object of class "ppll", usually, as a result of a call to
\code{prediction.profile.ll}.}

\item{...}{generic plot arguments.}
}
\value{
A xy-plot of one minus the cross-validation error (i.e. the
prediction accuracy against prediction time step.
}
\description{
`plot' method for class "ppll".
}
\details{
See \code{\link{prediction.profile.ll}} for details.
}
\seealso{
\code{\link{prediction.profile.ll}}
}
\keyword{ts}
