% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlts.R
\name{plot.specar.ci}
\alias{plot.specar.ci}
\title{Plot ar-spectra with CI's}
\usage{
\method{plot}{specar.ci}(x, period = TRUE, ...)
}
\arguments{
\item{x}{an object of class "specar.ci", usually, as a result of a call to
\code{\link{specar.ci}}.}

\item{period}{if TRUE x-axis is period, if FALSE frequency.}

\item{...}{generic plot arguments.}
}
\value{
A xy-plot of amplitude against period (or frequency).
}
\description{
`plot' method for class "specar.ci".
}
\seealso{
\code{\link{specar.ci}}
}
\keyword{ts}
