% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlts.R
\name{predict.ll.order}
\alias{predict.ll.order}
\title{Predict values from ll.order object.}
\usage{
\method{predict}{ll.order}(object, ...)
}
\arguments{
\item{object}{an object of class "ll.order", usually, as a result of a call
to \code{ll.order}.}

\item{\dots}{no other arguments currently allowed}
}
\value{
A data frame with observed and predicted values for the optimal
ll-model is returned.
}
\description{
Calculates the leave-one-out predicted values for the optimal ll.order
object
}
\details{
See \code{\link{ll.order}} for details.
}
\seealso{
\code{\link{ll.order}}
}
\keyword{ts}
