% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlts.R
\name{print.ll.order}
\alias{print.ll.order}
\title{Print nonparametric cross-validation for time-series order}
\usage{
\method{print}{ll.order}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{an object of class "ll.order", usually, as a result of a call to
\code{ll.order}.}

\item{verbose}{if TRUE provides a raw-printing of the object.}

\item{\dots}{no other arguments currently allowed}
}
\value{
A matrix summarizing the minimum cross-validation error (cv.min) and
the associated Gaussian-kernel bandwidth (bandwidth.opt) and model
degrees-of-freedom for each order considered.
}
\description{
`print' method for class "ll.order".
}
\details{
See \code{\link{ll.order}} for details.
}
\seealso{
\code{\link{ll.order}}
}
\keyword{ts}
