% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlts.R
\name{summary.lin.order}
\alias{summary.lin.order}
\title{Summarize linear cross-validation for time-series order}
\usage{
\method{summary}{lin.order}(object, ...)
}
\arguments{
\item{object}{an object of class "lin.order", usually, as a result of a call
to \code{lin.order.cls} or \code{lin.order.mle}.}

\item{\dots}{no other arguments currently allowed}
}
\value{
A slightly prettyfied version of the object is printed.
}
\description{
`summary' method for class "lin.order".
}
\seealso{
\code{\link{lin.order.cls}}
}
\keyword{ts}
