% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlts.R
\name{summary.ll.order}
\alias{summary.ll.order}
\title{Summarize nonparametric cross-validation for time-series order}
\usage{
\method{summary}{ll.order}(object, GCV = FALSE, ...)
}
\arguments{
\item{object}{an object of class "ll.order", usually, as a result of a call
to \code{ll.order}.}

\item{GCV}{if TRUE (or if cross-validation was not done), uses GCV values.}

\item{\dots}{no other arguments currently allowed}
}
\value{
A matrix summarizing the minimum cross-validation error (cv.min) and
the associated Gaussian-kernel bandwidth (bandwidth.opt) and model
degrees-of-freedom for each order considered.
}
\description{
`summary' method for class "ll.order".
}
\details{
See \code{\link{ll.order}} for details.
}
\seealso{
\code{\link{ll.order}}
}
\keyword{ts}
