% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlts.R
\name{summary.lomb}
\alias{summary.lomb}
\title{Summarizes Lomb periodograms}
\usage{
\method{summary}{lomb}(object, ...)
}
\arguments{
\item{object}{an object of class "lomb", usually, as a result of a call to
\code{spec.lomb}.}

\item{...}{generic plot arguments.}
}
\value{
A list summarizing the analysis is printed: \item{period}{the
dominant period.} \item{p.val}{the p.value.}
}
\description{
`summary' method for objects of class "lomb".
}
\seealso{
\code{\link{spec.lomb}}
}
\keyword{ts}
