% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlts.R
\name{summary.specar.ci}
\alias{summary.specar.ci}
\title{Summarize ar-spectra with CI's}
\usage{
\method{summary}{specar.ci}(object, period = TRUE, ...)
}
\arguments{
\item{object}{an object of class "specar.ci", usually, as a result of a call
to \code{specar.ci}.}

\item{period}{If TRUE the summary is given in terms of the period, if false
it is in terms of the frequency}

\item{...}{generic plot arguments.}
}
\value{
A list summarizing the analysis is printed: \item{period}{the
dominant period.} \item{p.val}{the p.value.}
}
\description{
`summary' method for objects of class "specar.ci".
}
\seealso{
\code{\link{specar.ci}}
}
\keyword{ts}
